Simplify solution evaluator
